Simplify logging interface by removing TaggedLogging wrapper
The `TaggedLogging` interface provided 2 things: 1) setting the formatter and 2) adding in support for the tagged interface into the logger. To simplify things, this leaves setting the formatter to the caller and moves the tagging interface directly into the logger.
