+++ lib/fluent_logger_rails.rb
@@ -1,5 +1,4 @@
 require 'fluent_logger_rails/logger'
-require 'fluent_logger_rails/tagged_logging'
 require 'fluent_logger_rails/hash_formatter'
 require 'fluent_logger_rails/json_formatter'
 require 'fluent_logger_rails/pretty_json_formatter'
